Move each strategy to separate files
