+++ TwitchChannelPointsMiner/classes/entities/strategies/HighOdds.py
@@ -1,0 +1,6 @@
+from TwitchChannelPointsMiner.classes.entities.Strategy import OutcomeKeys, Strategy
+
+
+class HighOdds(Strategy):
+    def calculate_middle(self):
+        self.decision["choice"] = self.return_choice(OutcomeKeys.ODDS)
