+++ TwitchChannelPointsMiner/classes/entities/strategies/MostVoted.py
@@ -1,0 +1,6 @@
+from TwitchChannelPointsMiner.classes.entities.Strategy import OutcomeKeys, Strategy
+
+
+class MostVoted(Strategy):
+    def calculate_middle(self):
+        self.decision["choice"] = self.return_choice(OutcomeKeys.TOTAL_USERS)
