+++ TwitchChannelPointsMiner/classes/entities/strategies/Percentage.py
@@ -1,0 +1,6 @@
+from TwitchChannelPointsMiner.classes.entities.Strategy import OutcomeKeys, Strategy
+
+
+class Percentage(Strategy):
+    def calculate_middle(self):
+        self.decision["choice"] = self.return_choice(OutcomeKeys.ODDS_PERCENTAGE)
